requirements.txt
python-dotenv
jinja2
requests